+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/http-request-action/utils/parseHttpJsonBodyWithoutVariablesOrThrow.ts
@@ -1,0 +1,5 @@
+import { removeVariablesFromJson } from '@/workflow/workflow-variables/utils/removeVariablesFromJson';
+
+export const parseHttpJsonBodyWithoutVariablesOrThrow = (value: string) => {
+  return JSON.parse(removeVariablesFromJson(value));
+};
